Home Work
Final Version for 30/09/2022

+++ factorial.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 
-int recurs(int n) {
+long long unsigned int recurs(int n) {
     if (n <= 1)
         return 1;
     return recurs(n - 1) * n;
@@ -10,7 +10,7 @@ int main(void) {
     int a;
     printf("Type the number of factorial you want to count : \n");
     scanf("%d", &a);
-    printf("Your factorial of %d is - > %d \n", a, recurs(a));
+    printf("Your factorial of %d is - > %llu \n", a, recurs(a));
 
     printf("To exit app type anything. \n");
