+++ fibraw.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 
-int fibonachi(int n) {
+long long unsigned int fibonachi(int n) {
     if (n <= 1)
         return n;
     return fibonachi(n - 2) + fibonachi(n - 1);
@@ -9,7 +9,7 @@ int fibonachi(int n) {
 int main (void) {
     printf("Type how many fibonachi number you want to out. \n"); 
     int array_size; scanf("%i", &array_size);
-    int a [array_size];
+    long long unsigned int a [array_size];
 
     printf("Your fibonachi raw : \n");
